Refactor compact method in Creature class to utilize CreatureUtil for improved clarity and efficiency. Removed unused imports and streamlined the logic for neuron and synapse management.
